Improve roo slope fixing, bgf load bug fix.
In the past, Meridian 59 slopes often incorporated vertexes from outside
the sector to give the desired plane properties, however this has the
disadvantage of making the slope impossible to copy/paste as
outside-sector vertex references are then invalid. Fully-working
copy/paste is useful enough to warrant limiting slopes to vertexes
inside the affected sector, and all 'invalid' vertex references must be
fixed.

This change attempts to recalculate broken slopes using specific
vertexes (1, 2, last or 1, last-1, last when ordered by height) to match
the original slope plane properties. Unfortunately this does not work
for every slope (e.g. East Tower in Ko'catan) as there won't always be
vertexes in the right places to give the desired slope angle, especially
with smaller sectors. Ultimately the solution is to move to the UDMF
slope format (extended Doom) but this requires changing both clients and
the old editor to match.

'Make Sectors' mode also attempts to fix slope properties on the old and
new sectors, if they were sloped.

Fixed a bug in BGF loading.