Fixes for sector slope checking, inc. epsilon for Vector3D.
Sector slope checking should account for sectors with no sidedefs as it
is called during 'Make Sector' mode which could end up with such a
sector.

Make the epsilon check in Vector3D larger, no difference on rendering
and allows more leeway on 'almost the same' values when fixing slopes in
old rooms.